+++ app/controllers/api/v1/tasks_controller.rb
@@ -1,0 +1,7 @@
+class Api::V1::TasksController < ApplicationController
+  def index
+  end
+
+  def create
+  end
+end
