+++ config/routes.rb
@@ -7,5 +7,5 @@ Rails.application.routes.draw do
       resources :skills
     end
   end
-  
+
 end
